requirements.txt
absl-py==2.3.1
asgiref==3.9.1
astunparse==1.6.3
beautifulsoup4==4.13.4
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.0
curl_cffi==0.12.0
cycler==0.12.1
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
flatbuffers==25.2.10
fonttools==4.58.5
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
multitasking==0.0.11
namex==0.1.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
protobuf==5.29.5
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.2.65
zipp==3.23.0
